Added working background worker

+++ app/src/main/java/es/alejandro/mtgspoileralert/settings/model/Settings.kt
@@ -2,7 +2,7 @@ package es.alejandro.mtgspoileralert.settings.model
 
 import java.util.concurrent.TimeUnit
 
-data class Settings (
+data class Settings(
     val coreSetListen: Boolean = false,
-    val interval: Pair<Int, TimeUnit> = Pair(15, TimeUnit.MINUTES)
+    val interval: Pair<Long, TimeUnit> = Pair(15, TimeUnit.MINUTES)
 )
